refactor and renaming of functions
