+++ tests/frontend/test_kibana_app/.gitignore
@@ -5,3 +5,4 @@
 node_modules/
 
 #env configurations
+cypress.env.json
