+++ tests/frontend/test_kibana_app/cypress/integration/pageobjects/wazuh-menu/rules.js
