+++ tests/frontend/test_kibana_app/cypress/integration/step-definitions/common/login-xpack.js
