+++ tests/frontend/test_kibana_app/cypress/integration/step-definitions/management/decoders/common/navigate-to-decoders.js
@@ -1,5 +1,5 @@
 import { When } from 'cypress-cucumber-preprocessor/steps';
-import WazuhMenu from '../../../pageobjects/wzMenu/wazuh-menu';
+import WazuhMenu from '../../../../pageobjects/wazuh-menu/wazuh-menu';
 
 When('The user navigates to decoders', () => {
   const wzMenu = new WazuhMenu();
