+++ tests/frontend/test_kibana_app/cypress/integration/step-definitions/management/decoders/shouldListCustomDecodersAndShowThePager.js
@@ -1,17 +1,23 @@
-import { Then } from 'cypress-cucumber-preprocessor/steps';
-import WazuhMenu from '../../../pageobjects/wzMenu/wazuh-menu';
+import { Then, When } from 'cypress-cucumber-preprocessor/steps';
+import WazuhMenu from '../../../pageobjects/wazuh-menu/wazuh-menu';
 
 const wzMenu = new WazuhMenu();
 const decoders = wzMenu.getDecoders();
 
-Then('The user should see the decoders', () => {
+When('The user press button custom decoders', () => {
   cy.wait(3000);
+  decoders.getCustomDecodersButton()
+    .click();
+});
 
+Then('The user should see the custom decoders', () => {
+  cy.wait(3000);
   decoders.getTittle()
     .should('exist')
     .should('contain', 'Decoders');
   decoders.getTable()
-    .should('exist');
+    .should('exist')
+    .should('be.visible');
   decoders.getTablePaginationDropdowns()
     .should('exist')
     .should('be.visible');
