+++ tests/frontend/test_kibana_app/cypress/integration/step-definitions/management/rules/shouldListRulesAndShowThePager.js
@@ -1,23 +1,15 @@
 import { Then } from 'cypress-cucumber-preprocessor/steps';
-import WazuhMenu from '../../../pageobjects/wzMenu/wazuh-menu';
+import WazuhMenu from '../../../pageobjects/wazuh-menu/wazuh-menu';
 
 const wzMenu = new WazuhMenu();
 const rules = wzMenu.getRules();
-
-Then('The user press button custom rules', () => {
-  cy.wait(3000);
-  rules.getCustomRulesButton()
-    .click();
-  cy.wait(3000);
-});
 
 Then('The user should see the rules', () => {
   rules.getTitle()
     .should('exist')
     .should('contain', 'Rules');
   rules.getTable()
-    .should('exist')
-    .should('be.visible');
+    .should('exist');
   rules.getDropdownPagination()
     .should('exist')
     .should('be.visible');
